video_video: Add the Virtio Video V4L2 driver
This adds a Virtio based video driver for video streaming device that
operates input and/or output data buffers to share video devices with
several guests. The current implementation consist of V4L2 based video
driver supporting video functions of decoder and encoder. This can be
extended for other functions e.g. processor, capture and output.
The device uses descriptor structures to advertise and negotiate stream
formats and controls. This allows the driver to modify the processing
logic of the device on a per stream basis.

Signed-off-by: Dmitry Sepp <Dmitry.Sepp@opensynergy.com>
Signed-off-by: Kiran Pawar <Kiran.Pawar@opensynergy.com>
Signed-off-by: Nikolay Martyanov <Nikolay.Martyanov@opensynergy.com>