Add custom qbuf and dqbuf ioctl implementations.
Currently some of the ioctls do not reach the driver because they are
handled in the v4l2 layer. This way the driver has no way to report a
generic stream error.

Add a simple wrapper implementations that first check the stream state
and then pass the parameters down the call stack.

Issue: COQOS-6916