CHROMIUM: drivers: media: virtio: Fix bug of re-registration of uuid
This invalidates a uuid associated with a vb2_buffer object when the
uuid is used for another vb2_buffer.

This patch will fix a problem that happens in scenarios like this:

(1) QBUF(index=0, DMABUF uuid=1)
(2) DQBUF(index=0)
(3) QBUF(index=1, DMABUF uuid=1)
(4) DQBUF(index=1)
(5) QBUF(index=0, DMABUF uuid=1)

Note that index is associated with a vb2_buffer and a resource_id here.

In this scenario, resource_create needs to be sent to register of a new
combination of resource_id and a uuid for (1), (3) and (5).

However, in the previous implementation, resource_create wasn't sent at (5)
because the combination (index=0, uuid=1) had been registered for (1) and
not been invalidated.

With this patch, a uuid in a vbb2_buffer object will be invalidated when
the uuid is tied with another vb2_buffer.
In this scenario, the combination registered at (1) will be invalidated
at (3).

Change-Id: Ie2eb7f50980cc2258540f692331a16aa11b5c334

[alexander.gordeev@opensynergy.com:
* Cherry picked from https://github.com/Gnurou/linux/commit/5d31ef447916.]

Issue: COQOS-17349