Remove the worker thread and optional job related callbacks.
After reviews it is clear that there is no real need for virtio-video to
postpone buffer submission and to handle it in a separate thread.

Submit buffers from buf_queue. Remove the worker thread. Remove optional job
callbacks as according to the current operation flow there is no concept of
jobs involved. The driver simply submits inputs and returns outputs.

Issue: COQOS-6992