Implement V4L2_CID_MIN_BUFFERS_FOR_OUTPUT control.
Gstreamer hardcodes min_buffers = 2 at the moment. But it gets a patch as well
in order to query the driver using this new control.

One or even two buffers are not enough in case of virtio-video. The pipeline
gets stalled every time because there are not enough buffers. Something like
this might happen for a decoder when it is parsing the encoded data, and the
buffer queued on OUTPUT is not strictly on a video frame boundary. This will
make the decoder wait for next buffer and not release the current buffer in
use. With a maximum of 2 buffers, it is observed that the Gstreamer plugin
only queues 1 buffer at a time.

Issue: COQOS-7097