CHROMIUM: drivers: media: virtio: propagate virtio errors
Treat VIRTIO_VIDEO_EVENT_ERROR as an unrecoverable errors for the
respective stream.

The spec isn't explicit as to whether or not the error event is fatal.
However, since the error event is generic, the cause is opaque to the
driver, so the driver can't meaningfully do anything to recover in the
case of a non-fatal error. Therefore it doesn't make sense for the
device to send non-fatal error events in the first place.

BUG=b:151703605, b:174445948
TEST=android.security.cts.StagefrightTest#testStagefright_bug_33818508

[5.4-arcvm: picked from 5.4]
Change-Id: I4d9eee47ff3e42058768c0e6e70c69518e8abe20
Signed-off-by: David Stevens <stevensd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2594587
Reviewed-by: Hikaru Nishida <hikalium@chromium.org>

[alexander.gordeev@opensynergy.com:
* Cherry picked from https://github.com/Gnurou/linux/commit/3746bc7f7438.]

Issue: COQOS-17349