Set maximum encoder bitrate to S32_MAX.
It was set to S64_MAX in commit 7bfd2f2, which was selected based on the size
of this parameter of v4l2_ctrl_new_std(). But it turns out this value is later
used as 32 bit signed integer. So S64_MAX turned to 0xffffffff or -1 later.
We still don't have a way to get the true maximum at the moment. Also hardware
drivers will limit the bitrate anyway. So set it to S32_MAX.

Issue: COQOS-9139