Use virtio_video_event structure for events.
Using of virtio_video_vbuffer structure is reduntant since it contains lot of
fields for flexible callbacks, requests and responses, which is necessary for
commands, but not needed for events. Actual evant data arrives from the host
with virtio_video_event structure for all event types.

Thus, in the updated approach we maintain static array of virtio_video_event
structures and resubmit them back on each response arrival.

Issue: COQOS-9453