media: virtio: respect V4L2_CID_MIN_BUFFERS_FOR_CAPTURE
The V4L2 m2m stateful specification says the following w.r.t VIDIOC_REQBUFS:

'The actual number of allocated buffers may differ from the count given. The
client must check the updated value of count after the call returns.'

The VIDIOC_REQBUFS documentation explicitly mentions that implementers are
free to allocate more buffers if needed, i.e.:

'A larger number is also possible when the driver requires more buffers to
function correctly.'

Thus, enforce a minimum number of capture buffers as per
V4L2_CID_MIN_BUFFERS_FOR_CAPTURE. This is very important in a zero-copy
workflow where the host is importing dma-bufs to serve as backing memory for
video data.

If the number of capture buffers allocated is not enough, it can stall or
corrupt the actual video decoder in the host, as codecs need a minimum number
of buffers to function. This is related to the DPB size, among other
codec-specific considerations.

Signed-off-by: Daniel Almeida <daniel.almeida@collabora.com>
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

[alexander.gordeev@opensynergy.com:
* Cherry picked from https://github.com/Gnurou/linux/commit/05f4d86a4869.]

Issue: COQOS-17349