Keep frame formats in their original order from the device.
The subsequent frame formats from the virtio messages were always
prepended to a local list, which reversed their order, and effectively
broke the mask indices. Out existing devices were not affected as
the mask was always all-ones. But in case the mask was not all-ones,
the VIDIOC_ENUM_FMT ioctl could return pixel formats not suitable
for current encoder/decoder.

Issue: COQOS-15787